Modified the model classes and set the title for the webpages
