+++ CryptoNewsAndroid/app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:theme="@style/AppTheme.NoActionBar" />
 
         <activity
-            android:name=".NewWebsite.view.ui.NewsWebPageActivity" >
+            android:name=".newswebsite.view.ui.NewsWebPageActivity" >
         <meta-data
             android:name="android.support.PARENT_ACTIVITY"
             android:value=".newslist.view.activity.LatestNewsActivity" />
