+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/api/NewsApiService.java
@@ -1,6 +1,6 @@
 package com.princess.android.cryptonews.api;
 
-import com.princess.android.cryptonews.model.ccn.News;
+import com.princess.android.cryptonews.model.News;
 
 import java.util.List;
