--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/Caption.java
@@ -1,27 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class Caption{
-
-	@SerializedName("rendered")
-	private String rendered;
-
-	public void setRendered(String rendered){
-		this.rendered = rendered;
-	}
-
-	public String getRendered(){
-		return rendered;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"Caption{" + 
-			"rendered = '" + rendered + '\'' + 
-			"}";
-		}
-}
