--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/CuriesItem.java
@@ -1,51 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class CuriesItem{
-
-	@SerializedName("templated")
-	private boolean templated;
-
-	@SerializedName("name")
-	private String name;
-
-	@SerializedName("href")
-	private String href;
-
-	public void setTemplated(boolean templated){
-		this.templated = templated;
-	}
-
-	public boolean isTemplated(){
-		return templated;
-	}
-
-	public void setName(String name){
-		this.name = name;
-	}
-
-	public String getName(){
-		return name;
-	}
-
-	public void setHref(String href){
-		this.href = href;
-	}
-
-	public String getHref(){
-		return href;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"CuriesItem{" + 
-			"templated = '" + templated + '\'' + 
-			",name = '" + name + '\'' + 
-			",href = '" + href + '\'' + 
-			"}";
-		}
-}
