--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/Excerpt.java
@@ -1,39 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class Excerpt{
-
-	@SerializedName("rendered")
-	private String rendered;
-
-	@SerializedName("protected")
-	private boolean jsonMemberProtected;
-
-	public void setRendered(String rendered){
-		this.rendered = rendered;
-	}
-
-	public String getRendered(){
-		return rendered;
-	}
-
-	public void setJsonMemberProtected(boolean jsonMemberProtected){
-		this.jsonMemberProtected = jsonMemberProtected;
-	}
-
-	public boolean isJsonMemberProtected(){
-		return jsonMemberProtected;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"Excerpt{" + 
-			"rendered = '" + rendered + '\'' + 
-			",protected = '" + jsonMemberProtected + '\'' + 
-			"}";
-		}
-}
