--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/RepliesItem.java
@@ -1,39 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class RepliesItem{
-
-	@SerializedName("href")
-	private String href;
-
-	@SerializedName("embeddable")
-	private boolean embeddable;
-
-	public void setHref(String href){
-		this.href = href;
-	}
-
-	public String getHref(){
-		return href;
-	}
-
-	public void setEmbeddable(boolean embeddable){
-		this.embeddable = embeddable;
-	}
-
-	public boolean isEmbeddable(){
-		return embeddable;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"RepliesItem{" + 
-			"href = '" + href + '\'' + 
-			",embeddable = '" + embeddable + '\'' + 
-			"}";
-		}
-}
