--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/Sizes.java
@@ -1,219 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class Sizes{
-
-	@SerializedName("w360")
-	private W360 w360;
-
-	@SerializedName("thumbnail")
-	private Thumbnail thumbnail;
-
-	@SerializedName("optin-monster-lightbox-theme-clean-slate")
-	private OptinMonsterLightboxThemeCleanSlate optinMonsterLightboxThemeCleanSlate;
-
-	@SerializedName("slider-image")
-	private SliderImage sliderImage;
-
-	@SerializedName("medium")
-	private Medium medium;
-
-	@SerializedName("w640")
-	private W640 w640;
-
-	@SerializedName("medium-thumb")
-	private MediumThumb mediumThumb;
-
-	@SerializedName("large-thumb")
-	private LargeThumb largeThumb;
-
-	@SerializedName("optin-monster-lightbox-theme-pendleton")
-	private OptinMonsterLightboxThemePendleton optinMonsterLightboxThemePendleton;
-
-	@SerializedName("trending-thumb")
-	private TrendingThumb trendingThumb;
-
-	@SerializedName("optin-monster-lightbox-theme-case-study")
-	private OptinMonsterLightboxThemeCaseStudy optinMonsterLightboxThemeCaseStudy;
-
-	@SerializedName("medium_large")
-	private MediumLarge mediumLarge;
-
-	@SerializedName("small-thumb")
-	private SmallThumb smallThumb;
-
-	@SerializedName("optin-monster-lightbox-theme-balance")
-	private OptinMonsterLightboxThemeBalance optinMonsterLightboxThemeBalance;
-
-	@SerializedName("optin-monster-lightbox-theme-bullseye")
-	private OptinMonsterLightboxThemeBullseye optinMonsterLightboxThemeBullseye;
-
-	@SerializedName("optin-monster-lightbox-theme-transparent")
-	private OptinMonsterLightboxThemeTransparent optinMonsterLightboxThemeTransparent;
-
-	@SerializedName("full")
-	private Full full;
-
-	public void setW360(W360 w360){
-		this.w360 = w360;
-	}
-
-	public W360 getW360(){
-		return w360;
-	}
-
-	public void setThumbnail(Thumbnail thumbnail){
-		this.thumbnail = thumbnail;
-	}
-
-	public Thumbnail getThumbnail(){
-		return thumbnail;
-	}
-
-	public void setOptinMonsterLightboxThemeCleanSlate(OptinMonsterLightboxThemeCleanSlate optinMonsterLightboxThemeCleanSlate){
-		this.optinMonsterLightboxThemeCleanSlate = optinMonsterLightboxThemeCleanSlate;
-	}
-
-	public OptinMonsterLightboxThemeCleanSlate getOptinMonsterLightboxThemeCleanSlate(){
-		return optinMonsterLightboxThemeCleanSlate;
-	}
-
-	public void setSliderImage(SliderImage sliderImage){
-		this.sliderImage = sliderImage;
-	}
-
-	public SliderImage getSliderImage(){
-		return sliderImage;
-	}
-
-	public void setMedium(Medium medium){
-		this.medium = medium;
-	}
-
-	public Medium getMedium(){
-		return medium;
-	}
-
-	public void setW640(W640 w640){
-		this.w640 = w640;
-	}
-
-	public W640 getW640(){
-		return w640;
-	}
-
-	public void setMediumThumb(MediumThumb mediumThumb){
-		this.mediumThumb = mediumThumb;
-	}
-
-	public MediumThumb getMediumThumb(){
-		return mediumThumb;
-	}
-
-	public void setLargeThumb(LargeThumb largeThumb){
-		this.largeThumb = largeThumb;
-	}
-
-	public LargeThumb getLargeThumb(){
-		return largeThumb;
-	}
-
-	public void setOptinMonsterLightboxThemePendleton(OptinMonsterLightboxThemePendleton optinMonsterLightboxThemePendleton){
-		this.optinMonsterLightboxThemePendleton = optinMonsterLightboxThemePendleton;
-	}
-
-	public OptinMonsterLightboxThemePendleton getOptinMonsterLightboxThemePendleton(){
-		return optinMonsterLightboxThemePendleton;
-	}
-
-	public void setTrendingThumb(TrendingThumb trendingThumb){
-		this.trendingThumb = trendingThumb;
-	}
-
-	public TrendingThumb getTrendingThumb(){
-		return trendingThumb;
-	}
-
-	public void setOptinMonsterLightboxThemeCaseStudy(OptinMonsterLightboxThemeCaseStudy optinMonsterLightboxThemeCaseStudy){
-		this.optinMonsterLightboxThemeCaseStudy = optinMonsterLightboxThemeCaseStudy;
-	}
-
-	public OptinMonsterLightboxThemeCaseStudy getOptinMonsterLightboxThemeCaseStudy(){
-		return optinMonsterLightboxThemeCaseStudy;
-	}
-
-	public void setMediumLarge(MediumLarge mediumLarge){
-		this.mediumLarge = mediumLarge;
-	}
-
-	public MediumLarge getMediumLarge(){
-		return mediumLarge;
-	}
-
-	public void setSmallThumb(SmallThumb smallThumb){
-		this.smallThumb = smallThumb;
-	}
-
-	public SmallThumb getSmallThumb(){
-		return smallThumb;
-	}
-
-	public void setOptinMonsterLightboxThemeBalance(OptinMonsterLightboxThemeBalance optinMonsterLightboxThemeBalance){
-		this.optinMonsterLightboxThemeBalance = optinMonsterLightboxThemeBalance;
-	}
-
-	public OptinMonsterLightboxThemeBalance getOptinMonsterLightboxThemeBalance(){
-		return optinMonsterLightboxThemeBalance;
-	}
-
-	public void setOptinMonsterLightboxThemeBullseye(OptinMonsterLightboxThemeBullseye optinMonsterLightboxThemeBullseye){
-		this.optinMonsterLightboxThemeBullseye = optinMonsterLightboxThemeBullseye;
-	}
-
-	public OptinMonsterLightboxThemeBullseye getOptinMonsterLightboxThemeBullseye(){
-		return optinMonsterLightboxThemeBullseye;
-	}
-
-	public void setOptinMonsterLightboxThemeTransparent(OptinMonsterLightboxThemeTransparent optinMonsterLightboxThemeTransparent){
-		this.optinMonsterLightboxThemeTransparent = optinMonsterLightboxThemeTransparent;
-	}
-
-	public OptinMonsterLightboxThemeTransparent getOptinMonsterLightboxThemeTransparent(){
-		return optinMonsterLightboxThemeTransparent;
-	}
-
-	public void setFull(Full full){
-		this.full = full;
-	}
-
-	public Full getFull(){
-		return full;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"Sizes{" + 
-			"w360 = '" + w360 + '\'' + 
-			",thumbnail = '" + thumbnail + '\'' + 
-			",optin-monster-lightbox-theme-clean-slate = '" + optinMonsterLightboxThemeCleanSlate + '\'' + 
-			",slider-image = '" + sliderImage + '\'' + 
-			",medium = '" + medium + '\'' + 
-			",w640 = '" + w640 + '\'' + 
-			",medium-thumb = '" + mediumThumb + '\'' + 
-			",large-thumb = '" + largeThumb + '\'' + 
-			",optin-monster-lightbox-theme-pendleton = '" + optinMonsterLightboxThemePendleton + '\'' + 
-			",trending-thumb = '" + trendingThumb + '\'' + 
-			",optin-monster-lightbox-theme-case-study = '" + optinMonsterLightboxThemeCaseStudy + '\'' + 
-			",medium_large = '" + mediumLarge + '\'' + 
-			",small-thumb = '" + smallThumb + '\'' + 
-			",optin-monster-lightbox-theme-balance = '" + optinMonsterLightboxThemeBalance + '\'' + 
-			",optin-monster-lightbox-theme-bullseye = '" + optinMonsterLightboxThemeBullseye + '\'' + 
-			",optin-monster-lightbox-theme-transparent = '" + optinMonsterLightboxThemeTransparent + '\'' + 
-			",full = '" + full + '\'' + 
-			"}";
-		}
-}
