--- CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/model/ccn/VersionHistoryItem.java
@@ -1,27 +1,0 @@
-package com.princess.android.cryptonews.model.ccn;
-
-import javax.annotation.Generated;
-import com.google.gson.annotations.SerializedName;
-
-@Generated("com.robohorse.robopojogenerator")
-public class VersionHistoryItem{
-
-	@SerializedName("href")
-	private String href;
-
-	public void setHref(String href){
-		this.href = href;
-	}
-
-	public String getHref(){
-		return href;
-	}
-
-	@Override
- 	public String toString(){
-		return 
-			"VersionHistoryItem{" + 
-			"href = '" + href + '\'' + 
-			"}";
-		}
-}
