+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/newslist/repository/NewsRepository.java
@@ -2,12 +2,11 @@ package com.princess.android.cryptonews.newslist.repository;
 
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.MutableLiveData;
-import android.util.Log;
 
 import com.princess.android.cryptonews.AppExecutors;
 import com.princess.android.cryptonews.api.NewsApiClient;
 import com.princess.android.cryptonews.api.NewsApiService;
-import com.princess.android.cryptonews.model.ccn.News;
+import com.princess.android.cryptonews.model.News;
 
 import java.util.List;
 
@@ -39,7 +38,6 @@ public class NewsRepository {
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(news ->{
                     newsMutableLiveData.setValue(news);
-                    //Log.e("TAG", news.toString());
                 });
         return newsMutableLiveData;
     }
