+++ CryptoNewsAndroid/app/src/main/java/com/princess/android/cryptonews/newswebsite/view/ui/NewsWebPageActivity.java
@@ -1,6 +1,5 @@
-package com.princess.android.cryptonews.NewWebsite.view.ui;
+package com.princess.android.cryptonews.newswebsite.view.ui;
 
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
@@ -8,7 +7,6 @@ import android.view.Menu;
 import android.view.MenuItem;
 
 import com.princess.android.cryptonews.R;
-import com.princess.android.cryptonews.settings.SettingsActivity;
 
 public class NewsWebPageActivity extends AppCompatActivity {
